+++ DeeptownOptimizator/HeightResource.h
@@ -3,8 +3,8 @@
 class Item;
 struct HeightResource
 {
-	const int numberPercentage;
+	const double numberPercentage;
 	const Item* const item;
-	HeightResource(int cNumberP, const Item* const cItem) :numberPercentage(cNumberP), item(cItem) {}
+	HeightResource(double cNumberP, const Item* const cItem) :numberPercentage(cNumberP), item(cItem) {}
 	double numberPercentageFraction() const { return numberPercentage * 0.01; }
 };
